+++ Client/FluentClientOptions.cs
@@ -6,10 +6,10 @@ namespace Pathoschild.Http.Client
         /*********
         ** Accessors
         *********/
-        /// <summary>Whether to ignore null arguments when the request is dispatched.</summary>
+        /// <summary>Whether to ignore null arguments when the request is dispatched. Default true if not specified.</summary>
         public bool? IgnoreNullArguments { get; set; }
 
-        /// <summary>Whether HTTP error responses (e.g. HTTP 404) should be ignored (else raised as exceptions).</summary>
+        /// <summary>Whether HTTP error responses (e.g. HTTP 404) should be ignored (else raised as exceptions). Default false if not specified.</summary>
         public bool? IgnoreHttpErrors { get; set; }